requirements.txt
opencv-python
streamlit
numpy
pillow
matplotlib
pyyaml